backend/requirements.txt
cryptography==41.0.3
Django==3.2.16
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.2.0
drf-extra-fields==3.7.0
djoser==2.2.0
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
pandas==2.1.1
numpy==1.26.0
sqlalchemy==2.0.21
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
